subscriber/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-pubsub